Add files via upload
Registration form which validates if the user has entered all the required data into form and after the validation is successful then it sends the data into table so that it can be displayed in tabular form also it check the duplicate entry. Further if the user wishes to edit an individual data then the user can do so by clicking on the data itself. And if the user wishes to edit all the data then it can be done by pressing edit under the option column also can delete the a particular row by clicking on delete button.